feat(mentorship): auto-embed mentor bio on acceptance and bio change
Extracts shared embedding helpers (embedBio, extractBioText,
shouldEmbedMentor) to src/lib/mentorship/embedBio.ts so both the
backfill script and the profile API share one implementation.

Profile PUT now fires a background embed when:
- mentor status flips to "accepted" (first-time embed), or
- mentor bio/about text changes while already accepted (re-embed)

Fire-and-forget — never blocks the API response. The backfill script
(scripts/embed-mentor-bios.ts) still needed for initial seed of
existing mentors.
